feat/40: add search in layer table, split component for pagination and search
