+++ app/package.json
@@ -23,7 +23,7 @@
     "cors": "^2.8.5",
     "crypto": "^1.0.1",
     "crypto-js": "^4.1.1",
-    "dotenv": "^16.0.3",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "html2canvas": "^1.4.1",
     "js-cookie": "^3.0.5",
