+++ app/server.js
@@ -1,4 +1,5 @@
 // server.js
+const workspaceAPI = require('./src/server/services/workspace')
 const projectionAPI = require('./src/server/projection')
 const featureAPI = require('./src/server/feature')
 const userAPI = require('./src/server/user')
@@ -71,6 +72,8 @@ app.get('/api/projections', projectionAPI.getAll)
 app.get('/api/projections/:id',projectionAPI.get)
 app.post('/api/projections', projectionAPI.create)
 app.put('/api/projections/:id', projectionAPI.update)
+// workspace
+app.get('/api/workspaces', workspaceAPI.getWorkspace)
 app.listen(process.env.PORT || 3000, () => {
   console.log(`Server started on port ${process.env.PORT || 3000}`)
 })
